+++ CHANGELOG.md
@@ -11,6 +11,7 @@ number or not.
 - [Revamped owner messages](https://basketball-gm.com/blog/2019/09/owner-messages-revamp/)
 - [All-Star Game](https://basketball-gm.com/blog/2019/09/all-star-game/)
 - New player faces, from faces.js v3.0.0
+- [More realistic play-by-plays in live game sim](https://basketball-gm.com/blog/2019/11/game-simulation-ovr-beta/)
 
 ## Technical details